wiseconnect: Import HAL version 4.1.1
Origin: WiSeConnect SDK
License: Zlib AND Apache-2.0
URL: https://github.com/SiliconLabs/wiseconnect
Commit: a0d475b5040e5126e23ee3c821d3728b8fdf57e5
Version: 4.1.1
Purpose: HAL for Silicon Labs SiWx91x devices

+++ wiseconnect/components/device/silabs/si91x/mcu/drivers/unified_peripheral_drivers/inc/sl_si91x_peripheral_gpio.h
@@ -70,19 +70,19 @@ extern "C" {
 #define _GPIO_PORT_K_PIN_MASK 0x0000UL
 
 ///< GPIO pins selection for selected port
-#define _GPIO_PORT_SIZE(port)              \
-  ((port) == 0    ? _GPIO_PORT_A_PIN_COUNT \
-   : (port) == 1  ? _GPIO_PORT_B_PIN_COUNT \
-   : (port) == 2  ? _GPIO_PORT_C_PIN_COUNT \
-   : (port) == 3  ? _GPIO_PORT_D_PIN_COUNT \
-   : (port) == 4  ? _GPIO_PORT_E_PIN_COUNT \
-   : (port) == 5  ? _GPIO_PORT_F_PIN_COUNT \
-   : (port) == 6  ? _GPIO_PORT_G_PIN_COUNT \
-   : (port) == 7  ? _GPIO_PORT_H_PIN_COUNT \
-   : (port) == 8  ? _GPIO_PORT_I_PIN_COUNT \
-   : (port) == 9  ? _GPIO_PORT_J_PIN_COUNT \
-   : (port) == 10 ? _GPIO_PORT_K_PIN_COUNT \
-                  : 0)
+#define _GPIO_PORT_SIZE(port)                              \
+  (((unsigned int)(port)) == 0    ? _GPIO_PORT_A_PIN_COUNT \
+   : ((unsigned int)(port)) == 1  ? _GPIO_PORT_B_PIN_COUNT \
+   : ((unsigned int)(port)) == 2  ? _GPIO_PORT_C_PIN_COUNT \
+   : ((unsigned int)(port)) == 3  ? _GPIO_PORT_D_PIN_COUNT \
+   : ((unsigned int)(port)) == 4  ? _GPIO_PORT_E_PIN_COUNT \
+   : ((unsigned int)(port)) == 5  ? _GPIO_PORT_F_PIN_COUNT \
+   : ((unsigned int)(port)) == 6  ? _GPIO_PORT_G_PIN_COUNT \
+   : ((unsigned int)(port)) == 7  ? _GPIO_PORT_H_PIN_COUNT \
+   : ((unsigned int)(port)) == 8  ? _GPIO_PORT_I_PIN_COUNT \
+   : ((unsigned int)(port)) == 9  ? _GPIO_PORT_J_PIN_COUNT \
+   : ((unsigned int)(port)) == 10 ? _GPIO_PORT_K_PIN_COUNT \
+                                  : 0)
 
 ///< GPIO pins mask for selected port
 #define _GPIO_PORT_MASK(port)                  \
@@ -327,7 +327,7 @@ static __INLINE void sl_gpio_set_pin_output(sl_gpio_port_t port, uint8_t pin)
   if ((port == SL_GPIO_PORT_A) || (port == SL_GPIO_PORT_B) || (port == SL_GPIO_PORT_C) || (port == SL_GPIO_PORT_D)) {
     SL_GPIO_ASSERT(SL_GPIO_NDEBUG_PORT_PIN(port, pin));
     GPIO->PIN_CONFIG[(port * MAX_GPIO_PORT_PIN) + pin].BIT_LOAD_REG = SET;
-  } else if (port == SL_GPIO_ULP_PORT) {
+  } else if ((unsigned int)port == SL_GPIO_ULP_PORT) {
     SL_GPIO_ASSERT(SL_GPIO_VALIDATE_ULP_PORT_PIN(port, pin));
     ULP_GPIO->PIN_CONFIG[pin].BIT_LOAD_REG = SET;
   } else {
@@ -355,7 +355,7 @@ static __INLINE void sl_gpio_set_pin_output(sl_gpio_port_t port, uint8_t pin)
 static __INLINE void sl_gpio_set_port_output(sl_gpio_port_t port, uint32_t pins)
 {
   SL_GPIO_ASSERT(SL_GPIO_VALIDATE_PORT(port));
-  if (port == SL_GPIO_ULP_PORT) {
+  if ((unsigned int)port == SL_GPIO_ULP_PORT) {
     ULP_GPIO->PORT_CONFIG[SL_GPIO_PORT_A].PORT_SET_REG = (pins);
   } else {
     GPIO->PORT_CONFIG[port].PORT_SET_REG = (pins);
@@ -381,7 +381,7 @@ static __INLINE void sl_gpio_set_port_output(sl_gpio_port_t port, uint32_t pins)
 static __INLINE void sl_gpio_set_port_output_value(sl_gpio_port_t port, uint32_t val, uint32_t mask)
 {
   SL_GPIO_ASSERT(SL_GPIO_VALIDATE_PORT(port));
-  if (port == SL_GPIO_ULP_PORT) {
+  if ((unsigned int)port == SL_GPIO_ULP_PORT) {
     ULP_GPIO->PORT_CONFIG[SL_GPIO_PORT_A].PORT_SET_REG = (ULP_GPIO->PORT_CONFIG[port].PORT_SET_REG & ~mask)
                                                          | (val & mask);
   } else {
@@ -429,7 +429,7 @@ static __INLINE void sl_gpio_clear_pin_output(sl_gpio_port_t port, uint8_t pin)
   if ((port == SL_GPIO_PORT_A) || (port == SL_GPIO_PORT_B) || (port == SL_GPIO_PORT_C) || (port == SL_GPIO_PORT_D)) {
     SL_GPIO_ASSERT(SL_GPIO_NDEBUG_PORT_PIN(port, pin));
     GPIO->PIN_CONFIG[(port * MAX_GPIO_PORT_PIN) + pin].BIT_LOAD_REG = CLR;
-  } else if (port == SL_GPIO_ULP_PORT) {
+  } else if ((unsigned int)port == SL_GPIO_ULP_PORT) {
     SL_GPIO_ASSERT(SL_GPIO_VALIDATE_ULP_PORT_PIN(port, pin));
     ULP_GPIO->PIN_CONFIG[pin].BIT_LOAD_REG = CLR;
   } else {
@@ -457,7 +457,7 @@ static __INLINE void sl_gpio_clear_pin_output(sl_gpio_port_t port, uint8_t pin)
 static __INLINE void sl_gpio_clear_port_output(sl_gpio_port_t port, uint32_t pins)
 {
   SL_GPIO_ASSERT(SL_GPIO_VALIDATE_PORT(port));
-  if (port == SL_GPIO_ULP_PORT) {
+  if ((unsigned int)port == SL_GPIO_ULP_PORT) {
     ULP_GPIO->PORT_CONFIG[SL_GPIO_PORT_A].PORT_CLEAR_REG = (pins);
   } else {
     GPIO->PORT_CONFIG[port].PORT_CLEAR_REG = (pins);
@@ -490,7 +490,7 @@ static __INLINE uint8_t sl_gpio_get_pin_input(sl_gpio_port_t port, uint8_t pin)
   if ((port == SL_GPIO_PORT_A) || (port == SL_GPIO_PORT_B) || (port == SL_GPIO_PORT_C) || (port == SL_GPIO_PORT_D)) {
     SL_GPIO_ASSERT(SL_GPIO_NDEBUG_PORT_PIN(port, pin));
     return (uint8_t)GPIO->PIN_CONFIG[(port * MAX_GPIO_PORT_PIN) + pin].BIT_LOAD_REG;
-  } else if (port == SL_GPIO_ULP_PORT) {
+  } else if ((unsigned int)port == SL_GPIO_ULP_PORT) {
     SL_GPIO_ASSERT(SL_GPIO_VALIDATE_ULP_PORT_PIN(port, pin));
     return (uint8_t)ULP_GPIO->PIN_CONFIG[pin].BIT_LOAD_REG;
   } else {
@@ -520,7 +520,7 @@ static __INLINE uint8_t sl_gpio_get_pin_input(sl_gpio_port_t port, uint8_t pin)
 static __INLINE uint8_t sl_gpio_get_pin_output(sl_gpio_port_t port, uint8_t pin)
 {
   SL_GPIO_ASSERT(SL_GPIO_VALIDATE_PORT(port));
-  if (port == SL_GPIO_ULP_PORT) {
+  if ((unsigned int)port == SL_GPIO_ULP_PORT) {
     SL_GPIO_ASSERT(SL_GPIO_VALIDATE_ULP_PORT_PIN(port, pin));
     return (uint8_t)ULP_GPIO->PIN_CONFIG[pin].BIT_LOAD_REG;
   } else {
@@ -546,7 +546,7 @@ static __INLINE uint8_t sl_gpio_get_pin_output(sl_gpio_port_t port, uint8_t pin)
 static __INLINE uint32_t sl_gpio_get_port_input(sl_gpio_port_t port)
 {
   SL_GPIO_ASSERT(SL_GPIO_VALIDATE_PORT(port));
-  if (port == SL_GPIO_ULP_PORT) {
+  if ((unsigned int)port == SL_GPIO_ULP_PORT) {
     return ULP_GPIO->PORT_CONFIG[SL_GPIO_PORT_A].PORT_READ_REG & PORT_MASK;
   } else {
     return GPIO->PORT_CONFIG[port].PORT_READ_REG & PORT_MASK;
@@ -570,7 +570,7 @@ static __INLINE uint32_t sl_gpio_get_port_input(sl_gpio_port_t port)
 static __INLINE uint32_t sl_gpio_get_port_output(sl_gpio_port_t port)
 {
   SL_GPIO_ASSERT(SL_GPIO_VALIDATE_PORT(port));
-  if (port == SL_GPIO_ULP_PORT) {
+  if ((unsigned int)port == SL_GPIO_ULP_PORT) {
     return (ULP_GPIO->PORT_CONFIG[SL_GPIO_PORT_A].PORT_READ_REG & PORT_MASK);
   } else {
     return (GPIO->PORT_CONFIG[port].PORT_READ_REG & PORT_MASK);
@@ -601,7 +601,7 @@ static __INLINE void sl_gpio_toggle_pin_output(sl_gpio_port_t port, uint8_t pin)
   if ((port == SL_GPIO_PORT_A) || (port == SL_GPIO_PORT_B) || (port == SL_GPIO_PORT_C) || (port == SL_GPIO_PORT_D)) {
     SL_GPIO_ASSERT(SL_GPIO_NDEBUG_PORT_PIN(port, pin));
     GPIO->PIN_CONFIG[(port * MAX_GPIO_PORT_PIN) + pin].BIT_LOAD_REG ^= SET;
-  } else if (port == SL_GPIO_ULP_PORT) {
+  } else if ((unsigned int)port == SL_GPIO_ULP_PORT) {
     SL_GPIO_ASSERT(SL_GPIO_VALIDATE_ULP_PORT_PIN(port, pin));
     ULP_GPIO->PIN_CONFIG[pin].BIT_LOAD_REG ^= SET;
   } else {
@@ -629,7 +629,7 @@ static __INLINE void sl_gpio_toggle_pin_output(sl_gpio_port_t port, uint8_t pin)
 static __INLINE void sl_gpio_toggle_port_output(sl_gpio_port_t port, uint32_t pins)
 {
   SL_GPIO_ASSERT(SL_GPIO_VALIDATE_PORT(port));
-  if (port == SL_GPIO_ULP_PORT) {
+  if ((unsigned int)port == SL_GPIO_ULP_PORT) {
     ULP_GPIO->PORT_CONFIG[SL_GPIO_PORT_A].PORT_TOGGLE_REG = pins;
   } else {
     GPIO->PORT_CONFIG[port].PORT_TOGGLE_REG = pins;
